+++ lib/orbis/nif.ex
@@ -86,6 +86,8 @@ defmodule Orbis.NIF do
 
   def sp3_satellite_ids(_handle), do: :erlang.nif_error(:nif_not_loaded)
 
+  def sp3_to_iodata(_handle), do: :erlang.nif_error(:nif_not_loaded)
+
   def broadcast_parse(_text), do: :erlang.nif_error(:nif_not_loaded)
 
   def broadcast_position(_handle, _system_letter, _prn, _t_j2000_s),